+++ src/idlib/hashing/Honeyman.h
@@ -2,6 +2,8 @@
 #ifndef __HONEYMAN_H__
 #define __HONEYMAN_H__
 
+#include <stdint.h>
+
 /*
 ===============================================================================
 
@@ -12,9 +14,9 @@
 ===============================================================================
 */
 
-void Honeyman_InitChecksum( unsigned long &crcvalue );
-void Honeyman_UpdateChecksum( unsigned long &crcvalue, const void *data, int length );
-void Honeyman_FinishChecksum( unsigned long &crcvalue );
-unsigned long Honeyman_BlockChecksum( const void *data, int length );
+void Honeyman_InitChecksum( uint32_t &crcvalue );
+void Honeyman_UpdateChecksum( uint32_t &crcvalue, const void *data, int length );
+void Honeyman_FinishChecksum( uint32_t &crcvalue );
+uint32_t Honeyman_BlockChecksum( const void *data, int length );
 
 #endif /* !__HONEYMAN_H__ */
