+++ src/idlib/hashing/MD4.h
@@ -2,6 +2,8 @@
 #ifndef __MD4_H__
 #define __MD4_H__
 
+#include <stdint.h>
+
 /*
 ===============================================================================
 
@@ -11,6 +13,6 @@
 ===============================================================================
 */
 
-unsigned long MD4_BlockChecksum( const void *data, int length );
+uint32_t MD4_BlockChecksum( const void *data, int length );
 
 #endif /* !__MD4_H__ */
